Add options to `amqp.publish` to allow binding do queues by default
RabbitMQ topic exchanges will drop messages if there are no subscribers
to a routing key. This means all consumers should be up before the
producer or we will lose messages. This is by design, since Rabbit won't
like to have a lot of queues that are just created and never consumed.

But for some special cases you'll want the queue to be created even if
the consumers are not up yet, so here we'll add the ability to do that.

To force a queue to be present, just send the `bind_to_queue` option to
the `#publish` method.